requirements.txt
PyPDF2>=3.0.0
anthropic>=0.7.0
python-dotenv>=1.0.0 